requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.4.26
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzlocal==5.3.1
